+++ daisygame/js/score.js
@@ -30,4 +30,21 @@ class Score {
   needToSave() {
     return this._prev_high_score < this._highScore;
   }
+
+  serialize() {
+    return {
+      score: this._score,
+      high: this._highScore,
+      prevHigh: this._prev_high_score,
+    };
+  }
+
+  restore(d) {
+    if (!d || typeof d !== 'object') return false;
+    if (typeof d.score !== 'number' || typeof d.high !== 'number') return false;
+    this._score = d.score;
+    this._highScore = d.high;
+    this._prev_high_score = typeof d.prevHigh === 'number' ? d.prevHigh : d.high;
+    return true;
+  }
 }